refactor - moving transaction history types to own file.
